+++ app/src/main/AndroidManifest.xml
@@ -42,7 +42,7 @@
         <!-- Recommended for Android 9 (API level 28) and lower. -->
         <!-- Required for Android 10 (API level 29) and higher. -->
         <service
-            android:name=".Location.LocationService"
+            android:name=".userlocation.LocationService"
             android:foregroundServiceType="location" >
         </service>
     </application>
