+++ app/src/main/java/com/bestapp/rice/userlocation/LocationService.kt
@@ -1,4 +1,4 @@
-package com.bestapp.rice.Location
+package com.bestapp.rice.userlocation
 
 import com.google.type.LatLng
 import kotlinx.coroutines.flow.Flow
